third_party/blink/web_tests/fullscreen/rendering/overflow-root.html

<!DOCTYPE html>
<title>overflow of root element when root element itself is fullscreen</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<style>:root { overflow: scroll; }</style>
<div id="log"></div>
<script>
async_test(function(t)
{
    var root = document.documentElement;
    trusted_request(t, root, document.body);

    document.addEventListener("fullscreenchange", t.step_func_done(function()
    {
        // The scrollbar should remain.
        assert_equals(getComputedStyle(root).overflow, "scroll");
        if (window.internals && internals.overlayScrollbarsEnabled)
          assert_equals(root.clientWidth, window.innerWidth);
        else
          assert_less_than(root.clientWidth, window.innerWidth);
    }));
});
</script>
